backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
gunicorn
openai
python-dotenv
